+++ frontend/src/App.tsx
@@ -1,61 +1,9 @@
-import { useCallback, useEffect, useState } from "react";
-import { ethers} from "ethers";
-import TokenArtifacts from './contracts/Token.json'
-import contractAddress from './contracts/contract-addresses.json'
-import { TokenData } from "./types/TokenData";
+import { useEffect } from "react";
 import { ConnectedView } from "./views/ConnectedView";
+import { useTokenContext } from './contexts/TokenContext';
 
 function App() {
-  const [connected, setConnected] = useState(false)
-  const [tokenData, setTokenData] = useState<TokenData>()
-
-  const connect = async () => {
-    if (connected) {
-      console.log("already connected")
-      return
-    } 
-
-    const provider = new ethers.providers.Web3Provider(window.ethereum);
-    await provider.send("eth_requestAccounts", []);
-    const signer = provider.getSigner()
-    const signerAddress = await signer.getAddress()
-    setConnected(true);
-  
-    const token = new ethers.Contract(contractAddress.Token, TokenArtifacts.abi, signer);
-    setTokenData(prevState => ({
-      ...prevState,
-      token,
-      signer,
-      signerAddress
-    }))
-  }
-
-  const refresh = useCallback( async () => {
-     
-    if (!tokenData?.token || !tokenData?.signer){
-      return;
-    }
-
-    console.log('refresh');
-    const { token, signer } = tokenData
-
-    const tokens = await token.balanceOf(signer.getAddress());
-
-    const tokenBalanceReadable = ethers.utils.formatUnits(tokens, 18);
-    const tokenName = await token.name();
-    const tokenSymbol = await token.symbol();
-    const isOwnerConnected = await token.isOwner();
-
-    setTokenData(prevState => ({
-      ...prevState,
-      tokenBalance: tokenBalanceReadable,
-      tokenName,
-      tokenSymbol,
-      isOwnerConnected,
-    }))
-  
-    
-  },[tokenData?.token, tokenData?.signer])
+  const {tokenData, connect, connected, refresh } = useTokenContext()
 
   useEffect(()=>{
     refresh()
@@ -66,7 +14,7 @@ function App() {
   }
 
   if (connected && tokenData) {
-    return <ConnectedView tokenData={tokenData} refresh={refresh} setTokenData={setTokenData}/>
+    return <ConnectedView />
   }
 
   return (
