+++ frontend/src/index.tsx
@@ -3,10 +3,13 @@ import ReactDOM from 'react-dom';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { TokenProvider } from './contexts/TokenContext';
 
 ReactDOM.render(
   <React.StrictMode>
-    <App />
+    <TokenProvider>
+      <App />
+    </TokenProvider>
   </React.StrictMode>,
   document.getElementById('root')
 );
